pyproject.toml
[tool.poetry]
name = "openai-gpt"
version = "0.1.0"
description = ""
authors = ["katewitkoski <ka.lwitkoski@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "0.28.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
